refactor: move files from `/reduxUtils` to `/entityOperations`
