+++ src/entityOperations/deleteEntitiesForTool.ts
