+++ src/modules/clients/sagas/clientsSagas.ts
@@ -2,17 +2,17 @@ import * as R from "ramda";
 import type { SagaIterator } from "redux-saga";
 import { call, put, select } from "redux-saga/effects";
 
-import { showErrorNotification } from "~/modules/app/appActions";
+import { linkEntitiesByIdByMapping } from "~/entityOperations/linkEntitiesByIdByMapping";
 import {
   toolActionSelector,
   toolNameByMappingSelector,
 } from "~/modules/allEntities/allEntitiesSelectors";
+import { showErrorNotification } from "~/modules/app/appActions";
 import * as clientsActions from "~/modules/clients/clientsActions";
 import {
   includedSourceClientsSelector,
   sourceClientsForTransferSelector,
 } from "~/modules/clients/clientsSelectors";
-import { linkEntitiesByIdByMapping } from "~/redux/reduxUtils";
 import { ClientsByIdModel, Mapping, ToolAction, ToolName } from "~/typeDefs";
 
 import * as clockifySagas from "./clockifyClientsSagas";
