+++ src/modules/projects/projectsSelectors.ts
@@ -2,13 +2,13 @@ import { createSelector } from "reselect";
 
 import * as R from "ramda";
 
+import { selectIdToLinkedId } from "~/entityOperations/selectIdToLinkedId";
 import {
   areExistsInTargetShownSelector,
   toolNameByMappingSelector,
 } from "~/modules/allEntities/allEntitiesSelectors";
 import { sourceTimeEntryCountByIdFieldSelectorFactory } from "~/modules/timeEntries/timeEntriesSelectors";
 import { activeWorkspaceIdSelector } from "~/modules/workspaces/workspacesSelectors";
-import { selectIdToLinkedId } from "~/redux/reduxUtils";
 import {
   ReduxState,
   ProjectModel,
