+++ src/modules/projects/sagas/projectsSagas.ts
@@ -3,11 +3,12 @@ import type { SagaIterator } from "redux-saga";
 import { all, call, put, select, takeEvery } from "redux-saga/effects";
 import { isActionOf } from "typesafe-actions";
 
-import { showErrorNotification } from "~/modules/app/appActions";
+import { linkEntitiesByIdByMapping } from "~/entityOperations/linkEntitiesByIdByMapping";
 import {
   toolActionSelector,
   toolNameByMappingSelector,
 } from "~/modules/allEntities/allEntitiesSelectors";
+import { showErrorNotification } from "~/modules/app/appActions";
 import * as projectActions from "~/modules/projects/projectsActions";
 import {
   includedSourceProjectsSelector,
@@ -16,7 +17,6 @@ import {
 } from "~/modules/projects/projectsSelectors";
 import { updateIsTaskIncluded } from "~/modules/tasks/tasksActions";
 import { sourceTasksSelector } from "~/modules/tasks/tasksSelectors";
-import { linkEntitiesByIdByMapping } from "~/redux/reduxUtils";
 import {
   Mapping,
   ProjectsByIdModel,
