+++ src/modules/tasks/tasksSelectors.ts
@@ -2,10 +2,10 @@ import { createSelector } from "reselect";
 
 import * as R from "ramda";
 
+import { selectIdToLinkedId } from "~/entityOperations/selectIdToLinkedId";
 import { sourceProjectsByIdSelector } from "~/modules/projects/projectsSelectors";
 import { sourceTimeEntryCountByIdFieldSelectorFactory } from "~/modules/timeEntries/timeEntriesSelectors";
 import { activeWorkspaceIdSelector } from "~/modules/workspaces/workspacesSelectors";
-import { selectIdToLinkedId } from "~/redux/reduxUtils";
 import {
   ReduxState,
   TaskModel,
