+++ src/modules/timeEntries/sagas/timeEntriesSagas.ts
@@ -2,17 +2,17 @@ import * as R from "ramda";
 import type { SagaIterator } from "redux-saga";
 import { call, put, select } from "redux-saga/effects";
 
-import { showErrorNotification } from "~/modules/app/appActions";
+import { linkEntitiesByIdByMapping } from "~/entityOperations/linkEntitiesByIdByMapping";
 import {
   toolActionSelector,
   toolNameByMappingSelector,
 } from "~/modules/allEntities/allEntitiesSelectors";
+import { showErrorNotification } from "~/modules/app/appActions";
 import * as timeEntriesActions from "~/modules/timeEntries/timeEntriesActions";
 import {
   includedSourceTimeEntriesSelector,
   sourceTimeEntriesForTransferSelector,
 } from "~/modules/timeEntries/timeEntriesSelectors";
-import { linkEntitiesByIdByMapping } from "~/redux/reduxUtils";
 import {
   Mapping,
   TimeEntriesByIdModel,
