+++ src/modules/workspaces/workspacesSelectors.ts
@@ -2,7 +2,7 @@ import { createSelector, createStructuredSelector } from "reselect";
 
 import * as R from "ramda";
 
-import { selectIdToLinkedId } from "~/redux/reduxUtils";
+import { selectIdToLinkedId } from "~/entityOperations/selectIdToLinkedId";
 import {
   Mapping,
   ReduxState,
